add auth

+++ backend/src/api.ts
@@ -4,8 +4,15 @@ import randomString from "crypto-random-string"
 import SHA512 from "crypto-js/sha512"
 import * as t from "io-ts"
 import StatusCodes from "http-status-codes"
-import { getUsers, DB, getLastUpdate, insertUser, deleteUser } from "./db"
-import { UserV } from "./data"
+import {
+  getUsers,
+  DB,
+  getLastUpdate,
+  insertUser,
+  deleteUser,
+  getUserByTokenhash,
+} from "./db"
+import { User, UserV } from "./data"
 import { decode } from "./io"
 
 export function getRoutes(db: DB) {
@@ -16,9 +23,11 @@ export function getRoutes(db: DB) {
     res.json(await getLastUpdate(db))
   })
   routes.get("/users", async (req: Request, res: Response) => {
+    await validateAdmin(req, db)
     res.json(await getUsers(db))
   })
   routes.post("/users", async (req: Request, res: Response) => {
+    await validateAdmin(req, db)
     const token = generateToken()
     const u = decode(UserV, {
       email: req.body.email,
@@ -29,13 +38,18 @@ export function getRoutes(db: DB) {
     res.json(token)
   })
   routes.delete("/users", async (req: Request, res: Response) => {
+    await validateAdmin(req, db)
     await deleteUser(db, decode(t.string, req.query.email))
     res.status(StatusCodes.NO_CONTENT).end()
   })
 
   // Errors
   routes.use((err: Error, req: Request, res: Response, _next: any) => {
-    res.status(StatusCodes.INTERNAL_SERVER_ERROR).json(err.message)
+    if (err.message.startsWith("UNAUTHORIZED")) {
+      res.status(StatusCodes.UNAUTHORIZED).json(err.message)
+    } else {
+      res.status(StatusCodes.INTERNAL_SERVER_ERROR).json(err.message)
+    }
   })
   return routes
 }
@@ -46,4 +60,22 @@ function generateToken(): string {
 
 function hash(s: string): string {
   return SHA512(s).toString()
+}
+
+async function validateUser(req: Request, db: DB): Promise<User> {
+  let token: string
+  try {
+    token = decode(t.string, req.header("Authorization")).split(" ")[1]
+  } catch (e) {
+    throw Error("UNAUTHORIZED: failed to parse auth header")
+  }
+  return await getUserByTokenhash(db, hash(token))
+}
+
+async function validateAdmin(req: Request, db: DB): Promise<User> {
+  const u = await validateUser(req, db)
+  if (u.accessGroup !== "admin") {
+    throw Error("UNAUTHORIZED: insufficient access")
+  }
+  return u
 }
